Move match processing code to processing_matches.py
now it's in its forever home <3
